Address PR review feedback
