+++ pkg/apis/core/v1beta1/helper/common_test.go
@@ -17,74 +17,71 @@ import (
 )
 
 var _ = Describe("Helper", func() {
-	Describe("Version helper", func() {
+	DescribeTable("#UpsertLastError",
+		func(lastErrors []gardencorev1beta1.LastError, lastError gardencorev1beta1.LastError, expected []gardencorev1beta1.LastError) {
+			Expect(UpsertLastError(lastErrors, lastError)).To(Equal(expected))
+		},
 
-		DescribeTable("#UpsertLastError",
-			func(lastErrors []gardencorev1beta1.LastError, lastError gardencorev1beta1.LastError, expected []gardencorev1beta1.LastError) {
-				Expect(UpsertLastError(lastErrors, lastError)).To(Equal(expected))
+		Entry(
+			"insert",
+			[]gardencorev1beta1.LastError{
+				{},
+				{TaskID: ptr.To("bar")},
 			},
-
-			Entry(
-				"insert",
-				[]gardencorev1beta1.LastError{
-					{},
-					{TaskID: ptr.To("bar")},
-				},
-				gardencorev1beta1.LastError{TaskID: ptr.To("foo"), Description: "error"},
-				[]gardencorev1beta1.LastError{
-					{},
-					{TaskID: ptr.To("bar")},
-					{TaskID: ptr.To("foo"), Description: "error"},
-				},
-			),
-			Entry(
-				"update",
-				[]gardencorev1beta1.LastError{
-					{},
-					{TaskID: ptr.To("foo"), Description: "error"},
-					{TaskID: ptr.To("bar")},
-				},
-				gardencorev1beta1.LastError{TaskID: ptr.To("foo"), Description: "new-error"},
-				[]gardencorev1beta1.LastError{
-					{},
-					{TaskID: ptr.To("foo"), Description: "new-error"},
-					{TaskID: ptr.To("bar")},
-				},
-			),
-		)
-
-		DescribeTable("#DeleteLastErrorByTaskID",
-			func(lastErrors []gardencorev1beta1.LastError, taskID string, expected []gardencorev1beta1.LastError) {
-				Expect(DeleteLastErrorByTaskID(lastErrors, taskID)).To(Equal(expected))
+			gardencorev1beta1.LastError{TaskID: ptr.To("foo"), Description: "error"},
+			[]gardencorev1beta1.LastError{
+				{},
+				{TaskID: ptr.To("bar")},
+				{TaskID: ptr.To("foo"), Description: "error"},
 			},
+		),
+		Entry(
+			"update",
+			[]gardencorev1beta1.LastError{
+				{},
+				{TaskID: ptr.To("foo"), Description: "error"},
+				{TaskID: ptr.To("bar")},
+			},
+			gardencorev1beta1.LastError{TaskID: ptr.To("foo"), Description: "new-error"},
+			[]gardencorev1beta1.LastError{
+				{},
+				{TaskID: ptr.To("foo"), Description: "new-error"},
+				{TaskID: ptr.To("bar")},
+			},
+		),
+	)
 
-			Entry(
-				"task id not found",
-				[]gardencorev1beta1.LastError{
-					{},
-					{TaskID: ptr.To("bar")},
-				},
-				"foo",
-				[]gardencorev1beta1.LastError{
-					{},
-					{TaskID: ptr.To("bar")},
-				},
-			),
-			Entry(
-				"task id found",
-				[]gardencorev1beta1.LastError{
-					{},
-					{TaskID: ptr.To("foo")},
-					{TaskID: ptr.To("bar")},
-				},
-				"foo",
-				[]gardencorev1beta1.LastError{
-					{},
-					{TaskID: ptr.To("bar")},
-				},
-			),
-		)
-	})
+	DescribeTable("#DeleteLastErrorByTaskID",
+		func(lastErrors []gardencorev1beta1.LastError, taskID string, expected []gardencorev1beta1.LastError) {
+			Expect(DeleteLastErrorByTaskID(lastErrors, taskID)).To(Equal(expected))
+		},
+
+		Entry(
+			"task id not found",
+			[]gardencorev1beta1.LastError{
+				{},
+				{TaskID: ptr.To("bar")},
+			},
+			"foo",
+			[]gardencorev1beta1.LastError{
+				{},
+				{TaskID: ptr.To("bar")},
+			},
+		),
+		Entry(
+			"task id found",
+			[]gardencorev1beta1.LastError{
+				{},
+				{TaskID: ptr.To("foo")},
+				{TaskID: ptr.To("bar")},
+			},
+			"foo",
+			[]gardencorev1beta1.LastError{
+				{},
+				{TaskID: ptr.To("bar")},
+			},
+		),
+	)
 
 	DescribeTable("#IsFailureToleranceTypeZone",
 		func(failureToleranceType *gardencorev1beta1.FailureToleranceType, expectedResult bool) {
@@ -124,37 +121,5 @@ var _ = Describe("Helper", func() {
 		Entry("migrate", metav1.ObjectMeta{Annotations: map[string]string{v1beta1constants.GardenerOperation: v1beta1constants.GardenerOperationMigrate}}, true),
 		Entry("unknown", metav1.ObjectMeta{Annotations: map[string]string{v1beta1constants.GardenerOperation: "unknown"}}, false),
 		Entry("not present", metav1.ObjectMeta{}, false),
-	)
-
-	DescribeTable("#AccessRestrictionsAreSupported",
-		func(seedAccessRestrictions []gardencorev1beta1.AccessRestriction, shootAccessRestrictions []gardencorev1beta1.AccessRestrictionWithOptions, expectation bool) {
-			Expect(AccessRestrictionsAreSupported(seedAccessRestrictions, shootAccessRestrictions)).To(Equal(expectation))
-		},
-
-		Entry("both have no access restrictions",
-			nil,
-			nil,
-			true,
-		),
-		Entry("shoot has no access restrictions",
-			[]gardencorev1beta1.AccessRestriction{{Name: "foo"}},
-			nil,
-			true,
-		),
-		Entry("seed has no access restrictions",
-			nil,
-			[]gardencorev1beta1.AccessRestrictionWithOptions{{AccessRestriction: gardencorev1beta1.AccessRestriction{Name: "foo"}}},
-			false,
-		),
-		Entry("both have access restrictions and they match",
-			[]gardencorev1beta1.AccessRestriction{{Name: "foo"}},
-			[]gardencorev1beta1.AccessRestrictionWithOptions{{AccessRestriction: gardencorev1beta1.AccessRestriction{Name: "foo"}}},
-			true,
-		),
-		Entry("both have access restrictions and they don't match",
-			[]gardencorev1beta1.AccessRestriction{{Name: "bar"}},
-			[]gardencorev1beta1.AccessRestrictionWithOptions{{AccessRestriction: gardencorev1beta1.AccessRestriction{Name: "foo"}}},
-			false,
-		),
 	)
 })
